Replace int by size_t in for loops.

+++ OndselSolver/Array.h
@@ -65,14 +65,14 @@ namespace MbD {
 	template<typename T>
 	inline void Array<T>::copyFrom(std::shared_ptr<Array<T>> x)
 	{
-		for (int i = 0; i < x->size(); i++) {
+		for (size_t i = 0; i < x->size(); i++) {
 			this->at(i) = x->at(i);
 		}
 	}
 	template<typename T>
 	inline void Array<T>::zeroSelf()
 	{
-		for (int i = 0; i < this->size(); i++) {
+		for (size_t i = 0; i < this->size(); i++) {
 			this->at(i) = (T)0;
 		}
 	}
@@ -97,7 +97,7 @@ namespace MbD {
 	//inline double Array<double>::maxMagnitude()
 	//{
 	//	double max = 0.0;
-	//	for (int i = 0; i < this->size(); i++)
+	//	for (size_t i = 0; i < this->size(); i++)
 	//	{
 	//		auto element = this->at(i);
 	//		if (element < 0.0) element = -element;
@@ -109,7 +109,7 @@ namespace MbD {
 	inline double Array<T>::maxMagnitudeOfVector()
 	{
 		double answer = 0.0;
-		for (int i = 0; i < this->size(); i++)
+		for (size_t i = 0; i < this->size(); i++)
 		{
 			double mag = std::abs(this->at(i));
 			if (answer < mag) answer = mag;
@@ -119,7 +119,7 @@ namespace MbD {
 	template<typename T>
 	inline void Array<T>::equalArrayAt(std::shared_ptr<Array<T>> array, int i)
 	{
-		for (int ii = 0; ii < this->size(); ii++)
+		for (size_t ii = 0; ii < this->size(); ii++)
 		{
 			this->at(ii) = array->at((size_t)i + ii);
 		}
@@ -141,7 +141,7 @@ namespace MbD {
 	//template<>
 	//inline void Array<double>::conditionSelfWithTol(double tol)
 	//{
-	//	for (int i = 0; i < this->size(); i++)
+	//	for (size_t i = 0; i < this->size(); i++)
 	//	{
 	//		double element = this->at(i);
 	//		if (element < 0.0) element = -element;
@@ -157,7 +157,7 @@ namespace MbD {
 	//inline double Array<double>::length()
 	//{
 	//	double ssq = 0.0;
-	//	for (int i = 0; i < this->size(); i++)
+	//	for (size_t i = 0; i < this->size(); i++)
 	//	{
 	//		double elem = this->at(i);
 	//		ssq += elem * elem;
@@ -167,7 +167,7 @@ namespace MbD {
 	template<typename T>
 	inline void Array<T>::magnifySelf(T factor)
 	{
-		for (int i = 0; i < this->size(); i++)
+		for (size_t i = 0; i < this->size(); i++)
 		{
 			this->atitimes(i, factor);
 		}
